bunch more cleanup
